ADDED: screenshots

+++ src/components/PostsList.jsx
@@ -1,23 +1,20 @@
 import { useState } from 'react';
 import {useGetPostsQuery, useAddPostMutation,useDeletePostMutation} from '../services/redux/apiSlice'
 import EditPost from './EditPost';
+import './styles/postList.css'
 
 const PostsList = () => { 
     const { data: posts, error, isLoading }  = useGetPostsQuery();
     const [title, setTitle] = useState('');
     const [body, setBody] = useState('');
     const [addPost] = useAddPostMutation();
-    const [deletePost] = useDeletePostMutation();
-
-    // State to manage the selected post to edit
+    const [deletePost] = useDeletePostMutation()
     const [editingPost, setEditingPost] = useState(null);
 
-    // Handler to set the selected post for editing
     const handleEditClick = (post) => {
         setEditingPost(post);
     };
 
-    // Handler to clear the selected post after editing
     const handleCancelEdit = () => {
         setEditingPost(null);
     };
@@ -32,7 +29,10 @@ const PostsList = () => {
       };
     
       const handleDelete = async (id) => {
+        const isConfirmed = window.confirm("Are you sure you want to delete this item?");
+        if (isConfirmed) {
         await deletePost(id);
+        }
       };
 
     if(isLoading) return <div>Loading....</div> 
@@ -41,19 +41,21 @@ const PostsList = () => {
     
     return (
         <div>
-          <h2>All Posts</h2>
+
+          { editingPost ? <h2>Edit the Post</h2> : <h2>All Posts</h2>}
           
-          {editingPost ? (
+          {
+          editingPost ? (
             <EditPost post={editingPost} onCancel={handleCancelEdit} />
           ) : ( 
             <> 
-            <form onSubmit={handleSubmit} > 
-            <input type="text" value={title} onChange={(e)=>setTitle(e.target.value)} placeholder='Title'/> 
-            <input type="text" value={body} onChange={(e)=>setBody(e.target.value)} placeholder='Content' style={{margin:20}}/> 
-            <button type='submit'>Add Post</button>
+            <form onSubmit={handleSubmit} className='form' > 
+              <div className='input-container'> 
+                <input type="text" value={title} onChange={(e)=>setTitle(e.target.value)} placeholder='Title'/> 
+                <input type="text" value={body} onChange={(e)=>setBody(e.target.value)} placeholder='Content' style={{margin:20}}/> 
+                <button type='submit'>Add Post</button>
+              </div>
             </form>
-            <br/>
-            <br/>
             <br/>
             {posts.map((post)=>{ 
                 const{id, title, body} = post 
@@ -61,12 +63,12 @@ const PostsList = () => {
                      <h3>{title}</h3>
                      <p>{body}</p>
                      <button onClick={() => handleDelete(id)}>Delete</button>
-                     <button onClick={() => handleEditClick(post)}>Edit</button>
+                     <button onClick={() => handleEditClick(post)} style={{marginLeft: '20px'}}>Edit</button>
                 </div>
-
             })}
             </>
-          )}
+          )
+          }
         </div>
       );
